Mainly UI
Can we change the colors on the custom token adding screen (pic below), its difficult to read the token name and symbol
DONE^

Can we changed the font to either Montserrat or Avenir. I think Montserrat may look better, but those are our font families.
DONE^

Can we get rid of the ending "#" in the url?
DONE^