Minor Changes
Create a new branch from the Main Branch, this has all the changes for Mainnet
	DONE^
Change the GUD page to remove the immediate connection request to Polygon and keep the page connected to Gauss. That means the page should default to swapping GUD to USDC. Once the users swaps that around, it should prompt then to connect to Polygon Mainnet.
	DONE^
Migrate the NobleSwap GUD page over to Gauss Mainnet and Polygon Mainnet
	DONE^
Migrate the OBurn fix to Gauss Mainnet
	MISSING CONTRACT ADDRESSES^